add hamberger icon, rename navbar to categoryBar
