+++ src/components/CatBar/CatBar.tsx
