+++ src/containers/CatBarContainer.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { NavBar } from "../components/NavBar/NavBar";
+import { NavBar } from "../components/CatBar/CatBar";
 
-interface NavBarContainerProp {
+interface CatBarContainerProp {
   onCategorySelected: (category: string) => void;
 }
-export const NavBarContainer: React.FC<NavBarContainerProp> = ({
+export const CatBarContainer: React.FC<CatBarContainerProp> = ({
   onCategorySelected,
 }) => {
   const categoriesArray = [
